refactor(app): split disconnect dialog into its own lazy module
openDisconnect imported dialog-connect-remote just to reach
DialogDisconnectRemote, dragging the whole pairing state machine (token input,
capture poll, spinner, toast) into the disconnect path's chunk. Move the
disconnect confirm into dialog-disconnect-remote.tsx so it lazy-loads only the
small confirm dialog; the connect flow is unchanged.
